requirements.txt
Flask
flask-cors
python-dotenv
ollama
qdrant-client
openai